refactor: SeisanCalculatorからDebtStateのコードを別ファイルに切り離した
